refactor(vue/system-status): break circular dep — shared types + post-Pinia wiring
Addresses three related PR comments — all rooted in the same anti-pattern
of treating the store as a de-facto types module:

  1. apiService.ts imported useSystemStatusStore (and ReadOnlyReasonCode)
     from stores/systemStatus, which itself imports apiService. ESM
     evaluation order is implementation-defined; partially-initialized
     exports could surface at runtime.
  2. models/websocket/systemStatusMessage.ts imported the same type
     from the store, coupling models/ to application state.
  3. The reason-code union + payload shape lived inside the Pinia
     store, making it a de-facto types module.

Two-part fix:

A. Shared types module — types/systemStatus.ts now exports
   ReadOnlyReasonCode and SystemStatusPayload. Every consumer (the
   store, apiService/interceptor, banner, stories, WS message model)
   imports types directly from there. No store re-export — these
   symbols are new in this PR and have no legacy call sites.

B. App-level interceptor wiring — moved the 503 read-only handler
   out of apiService.ts into a dedicated api/readOnlyInterceptor.ts
   that exports installReadOnlyInterceptor(). main.ts calls it
   AFTER `app.use(createPinia())`. apiService.ts is back to just the
   axios client + 401 handling and exports `apiClient` (named,
   alongside the existing default export of helper methods) so
   external installers can attach interceptors.

Resulting dep graph (no cycles, types as a shared leaf):

  types/systemStatus  ← stores, models, api/readOnlyInterceptor,
                        components/common/SystemStatusBanner(+story)
  api/apiService      ← stores/systemStatus, api/readOnlyInterceptor
  stores/systemStatus ← api/readOnlyInterceptor,
                        components/common/SystemStatusBanner(+story+spec)
  main.ts → api/readOnlyInterceptor (called after createPinia())

15/15 Vue tests still passing, vue-tsc clean, lint/prettier clean.
