+++ astros_vue/src/main.ts
@@ -21,6 +21,7 @@ import { MdDraghandle } from 'oh-vue-icons/icons/md';
 import App from './App.vue';
 import router from './router';
 import i18n from './i18n';
+import { installReadOnlyInterceptor } from '@/api/readOnlyInterceptor';
 
 addIcons(
   IoCloudUpload,
@@ -45,5 +46,11 @@ app.use(i18n);
 app.component('v-icon', OhVueIcon);
 app.use(createPinia());
 app.use(router);
+
+// Install the 503 read-only-mode response interceptor *after* Pinia is
+// active. This keeps apiService.ts free of any store import and breaks
+// what would otherwise be a circular module dependency
+// (apiService → store → apiService).
+installReadOnlyInterceptor();
 
 app.mount('#app');
